feat(TCOMP-1651): remove explicit Dynamic management
