+++ component-runtime-impl/src/main/java/org/talend/sdk/component/runtime/record/RecordConverters.java
@@ -76,7 +76,6 @@ import org.apache.johnzon.core.JsonLongImpl;
 import org.apache.johnzon.jsonb.extension.JsonValueReader;
 import org.talend.sdk.component.api.record.Record;
 import org.talend.sdk.component.api.record.Schema;
-import org.talend.sdk.component.api.record.dynamic.DynamicHelper;
 import org.talend.sdk.component.api.service.record.RecordBuilderFactory;
 import org.talend.sdk.component.runtime.record.json.OutputRecordHolder;
 import org.talend.sdk.component.runtime.record.json.PojoJsonbProvider;
@@ -509,11 +508,11 @@ public class RecordConverters implements Serializable {
 
     public interface IMappingMeta {
 
-        public boolean isLinearMapping();
+        boolean isLinearMapping();
 
-        public Object newInstance(final Record record);
+        Object newInstance(final Record record);
 
-        public <T> Record newRecord(final T data, final RecordBuilderFactory factory);
+        <T> Record newRecord(final T data, final RecordBuilderFactory factory);
     }
 
     @Data
@@ -783,7 +782,7 @@ public class RecordConverters implements Serializable {
                             instanceProvisioners.add(dynamicInstanceProvisioner.apply(field, name));
                             final Schema.Entry entry = builderFactory
                                     .newEntryBuilder()
-                                    .withName(name + DynamicHelper.DYNAMIC_MARKER)
+                                    .withName(name)
                                     .withRawName(name)
                                     .withType(RECORD)
                                     .withNullable(true)
